1.9.0-beta
Changelog
- Se han corregido errores menores:
1.-Se ha corregido el salto de linea en el login al introducir el usuario y contraseña.
2.-El campo solicitaba el email cuando se autenticaba con el usuario, se ha corregido el texto, ahora solicita correctamente el nombre de usuario.
3.-Se ha corregido un error que no permitía la introducción de decimales en la puntuación de los platos.
4.- Se ha corregido un bug que permitía que el filtro de solicitudes en la api pudiera falla
-Nuevas Funciones:
1.- Se ha añadido un buscador de restaurante a la hora de asignar el restaurante en un nuevo plato
2.- Se ha añadido una opción solo para administradores para asignar roles a diferentes  usuarios
3.- Se ha añadido la opción de editar los platos
4.- Se ha añadido solo para administradores la opción de editar los restaurantes.

+++ LoveThingsAPP/app/src/main/java/neoarcadia/core/lovethings/frames/MapsActivity.java
@@ -33,14 +33,12 @@ import retrofit2.Response;
 
 public class MapsActivity extends Fragment implements OnMapReadyCallback {
 
-
     private GoogleMap mMap;
     private ActivityMapsBinding binding;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.activity_maps, container, false);
-
 
         SupportMapFragment mapFragment = (SupportMapFragment) getChildFragmentManager()
                 .findFragmentById(R.id.mapfr);
@@ -75,7 +73,7 @@ public class MapsActivity extends Fragment implements OnMapReadyCallback {
                         if (!userDishes.isEmpty()) {
                             // Calcular nota media y platos pedidos
                             double averageRating = userDishes.stream()
-                                    .mapToInt(Dish::getRating)
+                                    .mapToDouble(Dish::getRating)
                                     .average()
                                     .orElse(0.0);
                             int dishesCount = userDishes.size();
